transformar id de Usuario a int (al menos de momento)

+++ yenny-app/src/domain/Usuario.java
@@ -1,21 +1,21 @@
 package domain;
 
 public class Usuario {
-    private Long id;
+    private int id;
     private String username;
     private String password;
     private Rol rol;
     private boolean activo = true;
 
-    public Usuario(Long id, String username, String password, Rol rol) {
+    public Usuario(int id, String username, String password, Rol rol) {
         this.id = id;
         this.username = username;
         this.password = password;
         this.rol = rol;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
